# Conflicts:
#	app/controllers/backend/DestinationController.java
#	app/controllers/backend/TripController.java
#	app/controllers/frontend/DestinationController.java
#	app/controllers/frontend/TripController.java
#	app/views/createTrip.scala.html
#	public/javascripts/destinations.js

+++ app/controllers/backend/DestinationController.java
@@ -4,15 +4,12 @@ import actions.ActionState;
 import actions.Authenticator;
 import actions.roles.Everyone;
 import com.fasterxml.jackson.databind.JsonNode;
-
 import java.io.IOException;
 import java.util.concurrent.CompletableFuture;
 import javax.inject.Inject;
-
 import models.Destination;
 import models.User;
 import play.libs.Json;
-import play.mvc.Controller;
 import play.mvc.Http;
 import play.mvc.Result;
 import play.mvc.With;
@@ -59,22 +56,24 @@ public class DestinationController extends TEABackController {
             newDestination.user.id = request.attrs().get(ActionState.USER).id;    // TODO: Modify to be whichever user its for (admin creating)
             return destinationRepository.addDestination(newDestination)
                 .thenApplyAsync(id -> {
-                    try{
+                    try {
                         return ok(sanitizeJson(Json.toJson(id)));
                     } catch (IOException e) {
-                        return internalServerError(Json.toJson("Sanitization Failed"));
+                        return internalServerError(Json.toJson(SANITIZATION_ERROR));
                     }
                 });
         }
     }
 
     /**
-     * Allows a user to mark one of their destinations as public, this will cause it to become immediately visible to all other users,
-     * as well as merging with any sufficiently similar destinations that are currently marked as private in the database
+     * Allows a user to mark one of their destinations as public, this will cause it to become
+     * immediately visible to all other users, as well as merging with any sufficiently similar
+     * destinations that are currently marked as private in the database
      *
      * @param request Request containing authentication header
      * @param id ID of destination to mark as public
-     * @return 200 if successful, 400 if already public, 401 unauthorized, 403 forbidden, 404 no such destination
+     * @return 200 if successful, 400 if already public, 401 unauthorized, 403 forbidden, 404 no
+     * such destination
      */
     @With({Everyone.class, Authenticator.class})
     public CompletableFuture<Result> makeDestinationPublic(Http.Request request, Long id) {
@@ -82,12 +81,14 @@ public class DestinationController extends TEABackController {
         // Try to get the destination, if it is not found throw 404
         return destinationRepository.getDestination(id).thenComposeAsync(destination -> {
             // Check for 404
-            if(destination == null) {
-                return CompletableFuture.supplyAsync(() -> notFound(Json.toJson("No such destination exists")));
+            if (destination == null) {
+                return CompletableFuture
+                    .supplyAsync(() -> notFound(Json.toJson("No such destination exists")));
             }
             // Check if user owns the destination (or is an admin)
-            if(!destination.user.id.equals(user.id) && !user.admin) {
-                return CompletableFuture.supplyAsync(() -> forbidden(Json.toJson("You are not allowed to perform this action")));
+            if (!destination.user.id.equals(user.id) && !user.admin) {
+                return CompletableFuture.supplyAsync(
+                    () -> forbidden(Json.toJson("You are not allowed to perform this action")));
             }
             // Otherwise perform the repository call which will return either 200, 400, or 404 as appropriate
             return destinationRepository.makeDestinationPublic(user, destination);
@@ -103,19 +104,65 @@ public class DestinationController extends TEABackController {
      * @return OK with number of rows deleted, badrequest if none deleted
      */
     @With({Everyone.class, Authenticator.class})
-    public CompletableFuture<Result> deleteDestination(Long id) {
-        // TODO: add authentication of user to destination
-        return destinationRepository.deleteDestination(id).thenApplyAsync(rowsDeleted -> {
-            if (rowsDeleted < 1) {
-                ErrorResponse errorResponse = new ErrorResponse();
-                errorResponse.map("Destination not found", "other");
-                return badRequest(errorResponse.toJson());
+    public CompletableFuture<Result> deleteDestination(Http.Request request, Long id) {
+        User user = request.attrs().get(ActionState.USER);
+        return destinationRepository.getDestination(id).thenComposeAsync(destination -> {
+            if (destination == null) {
+                return CompletableFuture.supplyAsync(() -> notFound("No such destination found"));
+            } else if (destination.user.id.equals(user.id) || user.admin) {
+                return destinationRepository.deleteDestination(id).thenApplyAsync(rowsDeleted -> {
+                    if (rowsDeleted < 1) {
+                        ErrorResponse errorResponse = new ErrorResponse();
+                        errorResponse.map("Destination not found", "other");
+                        return badRequest(errorResponse.toJson());
+                    } else {
+                        try {
+                            return ok(sanitizeJson(Json.toJson(rowsDeleted)));
+                        } catch (IOException e) {
+                            return internalServerError(Json.toJson(SANITIZATION_ERROR));
+                        }
+                    }
+                });
             } else {
-                try {
-                    return ok(sanitizeJson(Json.toJson(rowsDeleted)));
-                } catch (IOException e) {
-                    return internalServerError(Json.toJson("Sanitization Failed"));
+                return CompletableFuture.supplyAsync(() -> forbidden("Forbidden"));
+            }
+        });
+    }
+
+    /**
+     * Edits a destination's details with given id.
+     *
+     * @param request The request
+     * @param id The id of the destination to edit
+     * @return 400 is the request is bad, 404 if the destination is not found, 500 if sanitization
+     * fails, 403 if the user cannot edit the destination and 200 if successful
+     */
+    @With({Everyone.class, Authenticator.class})
+    public CompletableFuture<Result> editDestination(Http.Request request, Long id) {
+        JsonNode data = request.body().asJson();
+        User user = request.attrs().get(ActionState.USER);
+        ErrorResponse validatorResult = new DestinationValidator(data).addNewDestination();
+        return destinationRepository.getDestination(id).thenComposeAsync(destination -> {
+            if (destination == null) {
+                return CompletableFuture
+                    .supplyAsync(() -> notFound("Destination with provided ID not found"));
+            } else if (destination.user.id.equals(user.id) || user.admin) {
+                if (validatorResult.error()) {
+                    return CompletableFuture
+                        .supplyAsync(() -> badRequest(validatorResult.toJson()));
                 }
+                Destination editedDestination = Json.fromJson(data, Destination.class);
+                return destinationRepository.updateDestination(editedDestination)
+                    .thenApplyAsync(updatedDestination -> {
+                        try {
+                            return ok(sanitizeJson(Json.toJson("Successfully added destination")));
+                        } catch (IOException e) {
+                            return internalServerError(Json.toJson(SANITIZATION_ERROR));
+                        }
+                    });
+            } else {
+                return CompletableFuture.supplyAsync(() -> forbidden(
+                    "Forbidden, user does not have permission to edit this destination"));
             }
         });
     }
@@ -138,7 +185,7 @@ public class DestinationController extends TEABackController {
                         try {
                             return ok(sanitizeJson(Json.toJson(allDestinations)));
                         } catch (IOException e) {
-                            return internalServerError(Json.toJson("Sanitization Failed"));
+                            return internalServerError(Json.toJson(SANITIZATION_ERROR));
                         }
                     });
         }
@@ -149,7 +196,7 @@ public class DestinationController extends TEABackController {
                         try {
                             return ok(sanitizeJson(Json.toJson(allDestinations)));
                         } catch (IOException e) {
-                            return internalServerError(Json.toJson("Sanitization Failed"));
+                            return internalServerError(Json.toJson(SANITIZATION_ERROR));
                         }
                     });
         }
@@ -163,10 +210,10 @@ public class DestinationController extends TEABackController {
     public CompletableFuture<Result> getAllCountries() {
         return countryDefinitionRepository.getAllCountries()
             .thenApplyAsync(allCountries -> {
-                try{
+                try {
                     return ok(sanitizeJson(Json.toJson(allCountries)));
                 } catch (IOException e) {
-                    return internalServerError(Json.toJson("Sanitization Failed"));
+                    return internalServerError(Json.toJson(SANITIZATION_ERROR));
                 }
             });
     }
@@ -185,7 +232,7 @@ public class DestinationController extends TEABackController {
                 try {
                     return ok(sanitizeJson(Json.toJson(destination)));
                 } catch (IOException e) {
-                    return internalServerError(Json.toJson("Sanitization Failed"));
+                    return internalServerError(Json.toJson(SANITIZATION_ERROR));
                 }
             }
         });
@@ -201,7 +248,8 @@ public class DestinationController extends TEABackController {
      * @param filter The sort order (either asc or desc)
      * @return OK with paged list of destinations
      */
-    public CompletableFuture<Result> getPagedDestinations(int page, int pageSize, String order,
+    public CompletableFuture<Result> getPagedDestinations(int page, int pageSize, String
+        order,
         String filter) {
         // TODO: Destinations should be returned here which are not currently, update API spec when modified
         return destinationRepository.getPagedDestinations(page, pageSize, order, filter)
@@ -217,7 +265,12 @@ public class DestinationController extends TEABackController {
         return ok(
             JavaScriptReverseRouter.create("destinationRouter", "jQuery.ajax", request.host(),
                 controllers.backend.routes.javascript.DestinationController.getAllCountries(),
-                controllers.backend.routes.javascript.DestinationController.getAllDestinations()
+                controllers.backend.routes.javascript.DestinationController
+                    .getAllDestinations(),
+                controllers.backend.routes.javascript.DestinationController.getDestination(),
+                controllers.backend.routes.javascript.DestinationController.deleteDestination(),
+                controllers.frontend.routes.javascript.DestinationController
+                    .detailedDestinationIndex()
             )
         ).as(Http.MimeTypes.JAVASCRIPT);
     }
